requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.4
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
meraki==1.4.3
multidict==5.1.0
numpy==1.19.5
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
requests==2.25.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.4.0
